Rearrange code, remove duplicate.
