+++ src/main/java/uk/antiperson/stackmob/entity/EntityTools.java
@@ -1,23 +1,18 @@
 package uk.antiperson.stackmob.entity;
 
-import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.entity.*;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.metadata.FixedMetadataValue;
 
 import uk.antiperson.stackmob.StackMob;
 import uk.antiperson.stackmob.compat.PluginCompat;
 import uk.antiperson.stackmob.compat.hooks.MythicMobsHook;
-import uk.antiperson.stackmob.tools.GeneralTools;
-import uk.antiperson.stackmob.tools.extras.GlobalValues;
-
-import java.util.HashSet;
-import java.util.UUID;
 
 /**
  * Created by nathat on 24/07/17.
+ *
+ * Strictly stuff to do with entities, but not stacking.
  *
  * TODO: Seperate methods.
  */
@@ -32,20 +27,6 @@ public class EntityTools {
     // firstEntity and nearby are ALWAYS the same entity type!
     public boolean notMatching(Entity firstEntity, Entity nearby) {
         // Just checking if the entity is dead or not, just to be sure.
-        if(nearby.isDead()){
-            return true;
-        }
-
-        if(GeneralTools.hasValidMetadata(firstEntity, GlobalValues.NO_STACK_ALL) &&
-                firstEntity.getMetadata(GlobalValues.NO_STACK_ALL).get(0).asBoolean()){
-            return true;
-        }
-
-        if(GeneralTools.hasValidMetadata(nearby, GlobalValues.NO_STACK_ALL) &&
-                nearby.getMetadata(GlobalValues.NO_STACK_ALL).get(0).asBoolean()){
-            return true;
-        }
-
         if(sm.getHookManager().onEntityComparison(firstEntity, nearby)){
             return true;
         }
@@ -94,55 +75,15 @@ public class EntityTools {
     // TODO: fire ticks.
     public Entity cloneTraits(Entity original, Entity dupe){
         sm.getTraitManager().applyTraits(original, dupe);
-
-        // other plugin stuff
-        sm.hookManager.onEntityClone(dupe);
-
-        // noAi
-        setAi((LivingEntity) dupe);
-
+        setTraits((LivingEntity) dupe);
         return dupe;
     }
 
-    public boolean notTaskSuitable(Entity e){
-        if(e.isDead()){
-            return true;
-        }
-        return e.hasMetadata(GlobalValues.NO_STACK_ALL) && e.getMetadata(GlobalValues.NO_STACK_ALL).get(0).asBoolean();
-    }
-
-    public boolean notEnoughNearby(Entity original){
-        double xLoc = sm.config.getCustomConfig().getDouble("check-area.x");
-        double yLoc = sm.config.getCustomConfig().getDouble("check-area.y");
-        double zLoc = sm.config.getCustomConfig().getDouble("check-area.z");
-        if(sm.config.getCustomConfig().getInt("dont-stack-until") > 0){
-            HashSet<UUID> entities = new HashSet<>();
-            entities.add(original.getUniqueId());
-            for(Entity nearby : original.getNearbyEntities(xLoc, yLoc, zLoc)){
-                if(original.getType() == nearby.getType()) {
-                    if (GeneralTools.hasNotEnoughNear(nearby)) {
-                        if (notMatching(original, nearby)) {
-                            continue;
-                        }
-                        entities.add(nearby.getUniqueId());
-                    }
-                }
-            }
-            if(entities.size() >= sm.config.getCustomConfig().getInt("dont-stack-until")){
-                for(UUID uuid : entities){
-                    Entity nearby = Bukkit.getEntity(uuid);
-                    if(nearby == null){
-                        entities.remove(uuid);
-                        return true;
-                    }else{
-                        nearby.setMetadata(GlobalValues.METATAG, new FixedMetadataValue(sm, 1));
-                    }
-                }
-            }else{
-                return true;
-            }
-        }
-        return false;
+    public void setTraits(LivingEntity entity){
+        // other plugin stuff
+        sm.getHookManager().onEntityClone(entity);
+        // noAi
+        setAi(entity);
     }
 
     public void setAi(LivingEntity entity){
@@ -154,13 +95,5 @@ public class EntityTools {
             }
             entity.setAI(false);
         }
-    }
-
-    public boolean checkIfMaximumSize(Entity entity, int stackSize){
-        int maxStackSize = sm.getCustomConfig().getInt("stack-max");
-        if (sm.config.getCustomConfig().isInt("custom." + entity.getType() + ".stack-max")) {
-            maxStackSize =  sm.getCustomConfig().getInt("custom." + entity.getType() + ".stack-max");
-        }
-        return maxStackSize == stackSize;
     }
 }
